+++ CHANGELOG.md
@@ -6,8 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-
 ## [Unreleased]
+
+
+
+## [0.3.0] - 2026-09-20
+
+### Changed
+
+- Improve documentation.
 
 ### Removed
 
@@ -44,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Corrected beeswarm, quasirandom, and density behavior to match the upstream R implementations.
 
-[Unreleased]: https://github.com/ccwang002/p9beeswarm/compare/v0.2.1...HEAD
-[0.2.1]: https://github.com/ccwang002/p9beeswarm/releases/tag/v0.2.1
+[Unreleased]: https://github.com/ccwang002/p9beeswarm/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/ccwang002/p9beeswarm/compare/v0.2.1...v0.3.0
+[0.2.1]: https://github.com/ccwang002/p9beeswarm/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/ccwang002/p9beeswarm/releases/tag/v0.2.0